[GraphQL/MovePackage] Paginate by version
## Description

Introduce two new queries: `Query.packageVersions` and
`MovePackage.versions` for iterating over all the different versions
of a given package.

This kind of query is useful for understanding package history.

These were introduced as a separate query, instead of having a single
query for iterating over packages that could optionally take a
checkpoint bounds or version bounds because of how system packages
interact with the `packages` table:

Because system packages are updated in-place, they only have one row
in the `packages` table. This makes sense for paginating packages in
bulk (e.g. by checkpoint) where the primary aim is to get a snapshot
of the packages available at a certain point in time, but doesn't work
for answering package version queries for system packages, and it
prevents us from creating a combined query. A combined query would
also allow someone to create a filter that bounds checkpoints and
versions, but doesn't bound the package itself (or would require us to
prevent that combination), which is complicated to implement
efficiently and not particularly useful.

## Test plan

New E2E tests:

```
sui$ cargo nextest run -p sui-graphql-e2e-tests \
  --features pg_integration                     \
  -- packages/versioning
```

+ Testing against a read replica to make sure system package tests
work well, and performance is reasonable.